+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 <!-- version list -->
 
+## v0.2.0 (2026-09-25)
+
+### Features
+
+- Add average calculation logic
+  ([`1b9028e`](https://github.com/jaych415/jaych415-github-workshop/commit/1b9028e8125df3778136635cc98fef6a0aec2b7e))
+
+- **calculator**: Add reset button
+  ([`86bea4a`](https://github.com/jaych415/jaych415-github-workshop/commit/86bea4a238543c1c1630979b73d82f9503dab845))
+
+
 ## v0.1.0 (2026-09-25)
 
 - Initial Release